pyproject.toml
[tool.poetry]
name = "drf-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.1.3"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.10"
django-redis = "^5.4.0"
drf-spectacular = "^0.27.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
